Use collision sprite for boat
Becuase of the boats transparent windows pixel-perfect mouseover was not
working so add a fully white sprite in the shape of the boat and do
collision detection on that instead.

This would also allow masking if an object is behind another so that
mouse events do not trigger through the object in front.